refactor: Rename root directory to LoanSystemApi
