+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.nerd.LoanApi.LoanSystemApiApplication
+Main-Class: com.nerd.LoanSystemApi.LoanSystemApiApplication
