--- src/main/java/com/nerd/LoanApi/CustomException/LoanNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nerd.LoanApi.CustomException;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "A loan with provided id was not found.")
-public class LoanNotFoundException extends RuntimeException {
-}
