--- src/main/java/com/nerd/LoanApi/CustomException/UserNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nerd.LoanApi.CustomException;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "A user with the provided id was not found.")
-public class UserNotFoundException extends RuntimeException {
-}
