--- src/main/java/com/nerd/LoanApi/LoanSystemApiApplication.java
@@ -1,13 +1,0 @@
-package com.nerd.LoanApi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LoanSystemApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LoanSystemApiApplication.class, args);
-	}
-
-}
