--- src/main/java/com/nerd/LoanApi/model/contract/OrderBy.java
@@ -1,5 +1,0 @@
-package com.nerd.LoanApi.model.contract;
-
-public enum OrderBy {
-    NAME, INTEREST_RATE, OUTSTANDING_BALANCE, CONTRIBUTION
-}
