--- src/main/java/com/nerd/LoanApi/repository/UserDao.java
@@ -1,8 +1,0 @@
-package com.nerd.LoanApi.repository;
-
-import com.nerd.LoanApi.model.provider.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserDao extends JpaRepository<User, Integer> {
-
-}
