+++ src/main/java/com/nerd/LoanSystemApi/CustomException/UserNotFoundException.java
@@ -1,0 +1,8 @@
+package com.nerd.LoanSystemApi.CustomException;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "A user with the provided id was not found.")
+public class UserNotFoundException extends RuntimeException {
+}
