+++ src/main/java/com/nerd/LoanSystemApi/LoanSystemApiApplication.java
@@ -1,0 +1,13 @@
+package com.nerd.LoanSystemApi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LoanSystemApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LoanSystemApiApplication.class, args);
+	}
+
+}
