+++ src/main/java/com/nerd/LoanSystemApi/model/contract/OrderBy.java
@@ -1,0 +1,5 @@
+package com.nerd.LoanSystemApi.model.contract;
+
+public enum OrderBy {
+    NAME, INTEREST_RATE, OUTSTANDING_BALANCE, CONTRIBUTION
+}
