+++ src/main/java/com/nerd/LoanSystemApi/model/provider/User.java
@@ -1,0 +1,36 @@
+package com.nerd.LoanSystemApi.model.provider;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+
+@NoArgsConstructor
+@Entity
+@Getter @Setter @EqualsAndHashCode(onlyExplicitlyIncluded = true)
+@Table(name = "users")
+public class User implements Serializable {
+
+    @EqualsAndHashCode.Include
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Integer id;
+
+    @Column(name = "first_name")
+    private String firstName;
+
+    @Column(name = "last_name")
+    private String lastName;
+
+    @OneToMany(
+            mappedBy = "user",
+            cascade = CascadeType.ALL,
+            orphanRemoval = true
+    )
+    List<Loan> loans = new ArrayList<>();
+}
