+++ src/main/java/com/nerd/LoanSystemApi/repository/UserDao.java
@@ -1,0 +1,8 @@
+package com.nerd.LoanSystemApi.repository;
+
+import com.nerd.LoanSystemApi.model.provider.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserDao extends JpaRepository<User, Integer> {
+
+}
