--- src/test/java/com/nerd/LoanApi/LoanSystemApiApplicationTests.java
@@ -1,16 +1,0 @@
-package com.nerd.LoanApi;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.junit.MockitoJUnitRunner;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@RunWith(MockitoJUnitRunner.class)
-@SpringBootTest
-public class LoanSystemApiApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
