+++ src/test/java/com/nerd/LoanSystemApi/LoanSystemApiApplicationTests.java
@@ -1,0 +1,16 @@
+package com.nerd.LoanSystemApi;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@RunWith(MockitoJUnitRunner.class)
+@SpringBootTest
+public class LoanSystemApiApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
